Test that a character made with initiative 10 actually has initiative 10, and the same for a character with initiative 20

+++ src/main/java/framework/Charact.java
@@ -15,7 +15,7 @@ public interface Charact {
      * Returns the initiative of the character.
      * @return The initiative of the character.
      */
-    double getInitiative();
+    int getInitiative();
 
     /**
      * Returns the name of the character.
